+++ stdnet/contrib/tasks/queue.py
@@ -1,3 +1,4 @@
+from multiprocessing.queues import Empty
 
 from stdnet import orm
 
@@ -15,21 +16,32 @@ as a python multiprocessing Queue. To use it::
     name = orm.CharField(required = False)
     queue = orm.ListField()
     
+    def __unicode__(self):
+        return self.name
+    
     def put(self, elem):
         '''Put item into the queue.'''
         q = self.queue
         q.push_front(elem)
         q.save()
         
-    def get(self, block = True, timeout = 0):
-        '''Remove and return an item from the queue. If optional args ``block`` is ``True`` (the default)
-and ``timeout`` is ``0`` (the default), block if necessary until an item is available.
-If timeout is a positive number, it blocks at most ``timeout`` seconds and 
-return ``None`` if no item was available within that time.'''
+    def get(self, block = True, timeout = None):
+        '''Remove and return an item from the queue.
+If optional args ``block`` is ``True`` (the default)
+and ``timeout`` is ``0`` (the default), block if necessary until
+an item is available.
+If timeout is a positive number, it blocks at most ``timeout`` seconds.
+
+It raises  and `multiprocessing.queues.Empty` exception
+if no item was available.'''
         if block:
-            return self.queue.block_pop_back(timeout)
+            timeout = max(1,int(round(timeout))) if timeout else 0
+            ret = self.queue.block_pop_back(timeout)
         else:
-            return self.queue.pop_back()
+            ret = self.queue.pop_back()
+        if ret is None:
+            raise Empty
+        return ret
         
     def qsize(self):
         '''Return the approximate size of the queue.'''