+++ stdnet/contrib/tasks/tests/regression.py
@@ -4,7 +4,9 @@ from random import random
 
 from stdnet import test
 from stdnet.utils import populate
-from stdnet.contrib import tasks 
+from stdnet.contrib import tasks
+
+from .models import TaskQueue, Task
 
 
 TEST_LEN = 100
@@ -13,19 +15,19 @@ names = populate('string',TEST_LEN, min_len = 5, max_len = 10)
 
 def timeit(f, *args, **kwargs):
     t = time()
-    v = f(*args,**kwargs)
+    try:
+        v = f(*args,**kwargs)
+    except tasks.Empty:
+        v = None
     t = time() - t
     return v,t
 
 
-class TestQueue(test.TestCase):
+class TestQueue(test.TestModelBase):
+    model = tasks.Queue
     
     def setUp(self):
-        self.orm.register(tasks.Queue)
-        self.q = tasks.Queue(name = 'test').save()
-        
-    def unregister(self):
-        self.orm.unregister(tasks.Queue)
+        self.q = self.model(name = 'test').save()
         
     def testPutGet(self):
         N = len(names)
@@ -47,7 +49,7 @@ class TestQueue(test.TestCase):
         self.assertEqual(v,None)
         self.assertTrue(t<1)
         
-    def __testTimeout(self):
+    def testTimeout(self):
         
         def putData(q):
             for i in range(10):
@@ -61,7 +63,21 @@ class TestQueue(test.TestCase):
             
         t1 = threading.Thread(target=putData, args = (self.q,))
         t2 = threading.Thread(target=getData, args = (self.q,))
-        
         t1.start()
         t2.start()
+        t1.join()
+        t2.join()
         
+        
+class TestTaskQueue(test.TestModelBase):
+    models = (TaskQueue,Task)
+    
+    def setUp(self):
+        self.q = TaskQueue(name = 'test').save()
+        
+    def testPutGet(self):
+        q = self.q
+        q.put(Task(name = test).save())
+        self.assertEqual(q.qsize(),1)
+        t = q.get()
+        self.assertTrue(isinstance(t,Task))